Move the core plugins to their own file
